+++ addons/mail/models/__init__.py
@@ -15,6 +15,8 @@ from . import mail_thread
 from . import mail_thread_blacklist
 from . import mail_thread_cc
 from . import mail_blacklist
+from . import mail_moderation
+from . import mail_channel_partner
 from . import mail_channel
 from . import mail_template
 from . import mail_shortcode
